# docs/swagger/swagger.yaml
basePath: /
definitions:
  controller.KlingCameraConfig:
    properties:
      horizontal:
        example: 2.5
        type: number
      pan:
        example: 0
        type: number
      roll:
        example: 0
        type: number
      tilt:
        example: 0
        type: number
      vertical:
        example: 0
        type: number
      zoom:
        example: 0
        type: number
    type: object
  controller.KlingCameraControl:
    properties:
      config:
        $ref: '#/definitions/controller.KlingCameraConfig'
      type:
        example: simple
        type: string
    type: object
  controller.KlingImage2VideoRequest:
    properties:
      aspect_ratio:
        example: "16:9"
        type: string
      callback_url:
        example: https://your.domain/callback
        type: string
      camera_control:
        $ref: '#/definitions/controller.KlingCameraControl'
      cfg_scale:
        example: 0.7
        type: number
      duration:
        example: "5"
        type: string
      external_task_id:
        example: custom-task-002
        type: string
      image:
        example: https://h2.inkwai.com/bs2/upload-ylab-stunt/se/ai_portal_queue_mmu_image_upscale_aiweb/3214b798-e1b4-4b00-b7af-72b5b0417420_raw_image_0.jpg
        type: string
      mode:
        example: std
        type: string
      model_name:
        example: kling-v2-master
        type: string
      negative_prompt:
        example: blurry, low quality
        type: string
      prompt:
        example: A cat playing piano in the garden
        type: string
    required:
    - image
    type: object
  controller.KlingText2VideoRequest:
    properties:
      aspect_ratio:
        example: "16:9"
        type: string
      callback_url:
        example: https://your.domain/callback
        type: string
      camera_control:
        $ref: '#/definitions/controller.KlingCameraControl'
      cfg_scale:
        example: 0.7
        type: number
      duration:
        example: "5"
        type: string
      external_task_id:
        example: custom-task-001
        type: string
      mode:
        example: std
        type: string
      model_name:
        example: kling-v1
        type: string
      negative_prompt:
        example: blurry, low quality
        type: string
      prompt:
        example: A cat playing piano in the garden
        type: string
    required:
    - prompt
    type: object
  controller.VideoErrorPayload:
    properties:
      message:
        example: task_id is required
        type: string
      type:
        example: invalid_request_error
        type: string
    type: object
  controller.VideoErrorResponse:
    properties:
      error:
        $ref: '#/definitions/controller.VideoErrorPayload'
    type: object
  controller.VideoGenerationsRequest:
    properties:
      audios:
        description: 参考音频 URL 列表（仅 AudioVideoRef 模型生效）
        items:
          type: string
        type: array
      duration:
        example: 5
        type: integer
      image:
        description: 兼容单图；images 为空时会自动复制到 images[0]
        type: string
      images:
        description: 参考图 URL 列表
        items:
          type: string
        type: array
      metadata:
        additionalProperties: {}
        description: 透传给上游渠道的扩展字段
        type: object
      mode:
        description: 'Doubao: first_frame / first_last_frame / multi_reference'
        example: multi_reference
        type: string
      model:
        example: doubao-seedance-2-0-260128
        type: string
      prompt:
        example: 宇航员在月面慢跑，背景是地球
        type: string
      seconds:
        description: 字符串秒数；>0 时覆盖 duration
        example: "5"
        type: string
      size:
        example: 720x1280
        type: string
      videos:
        description: 参考视频 URL 列表（仅 AudioVideoRef 模型生效）
        items:
          type: string
        type: array
    required:
    - prompt
    type: object
  dto.OpenAIVideo:
    properties:
      completed_at:
        type: integer
      created_at:
        type: integer
      error:
        $ref: '#/definitions/dto.OpenAIVideoError'
      expires_at:
        type: integer
      id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      model:
        type: string
      object:
        type: string
      progress:
        type: integer
      remixed_from_video_id:
        type: string
      seconds:
        type: string
      size:
        type: string
      status:
        description: 'Should use VideoStatus constants: VideoStatusQueued, VideoStatusInProgress,
          VideoStatusCompleted, VideoStatusFailed'
        type: string
      task_id:
        description: 兼容旧接口 待废弃
        type: string
    type: object
  dto.OpenAIVideoError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  dto.TaskDto:
    properties:
      action:
        type: string
      channel_id:
        type: integer
      created_at:
        type: integer
      data:
        items:
          type: integer
        type: array
      fail_reason:
        type: string
      finish_time:
        type: integer
      group:
        type: string
      id:
        type: integer
      platform:
        type: string
      progress:
        type: string
      properties: {}
      quota:
        type: integer
      result_url:
        description: 任务结果 URL（视频地址等）
        type: string
      start_time:
        type: integer
      status:
        type: string
      submit_time:
        type: integer
      task_id:
        type: string
      updated_at:
        type: integer
      user_id:
        type: integer
      username:
        type: string
    type: object
  dto.VideoTaskError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  dto.VideoTaskMetadata:
    properties:
      duration:
        description: 实际生成的视频时长
        example: 5
        type: number
      fps:
        description: 实际帧率
        example: 30
        type: integer
      height:
        description: 实际高度
        example: 512
        type: integer
      seed:
        description: 使用的随机种子
        example: 20231234
        type: integer
      width:
        description: 实际宽度
        example: 512
        type: integer
    type: object
  dto.VideoTaskResponse:
    properties:
      error:
        allOf:
        - $ref: '#/definitions/dto.VideoTaskError'
        description: 错误信息（失败时）
      format:
        description: 视频格式
        example: mp4
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/dto.VideoTaskMetadata'
        description: 结果元数据
      status:
        description: 任务状态
        example: succeeded
        type: string
      task_id:
        description: 任务ID
        example: abcd1234efgh
        type: string
      url:
        description: 视频资源URL（成功时）
        type: string
    type: object
info:
  contact: {}
  description: |-
    New API 网关的 OpenAPI 文档（当前仅覆盖 /video 相关路由）。
    完整项目地址：https://github.com/QuantumNous/new-api
  title: New API
  version: "1.0"
paths:
  /jimeng/:
    post:
      consumes:
      - application/json
      description: |-
        即梦官方 API 统一入口，通过 query 参数 `Action` + `Version` 路由到提交或查询动作。
        - `Action=CVSync2AsyncSubmitTask&Version=2022-08-31` 提交任务
        - `Action=CVSync2AsyncGetResult&Version=2022-08-31` 查询结果
        参考：https://www.volcengine.com/docs/85621/1538636
      parameters:
      - description: 操作类型
        enum:
        - CVSync2AsyncSubmitTask
        - CVSync2AsyncGetResult
        in: query
        name: Action
        required: true
        type: string
      - description: API 版本
        example: "2022-08-31"
        in: query
        name: Version
        required: true
        type: string
      - description: 即梦原生请求体（字段视 Action 而定）
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 任务状态和结果
          schema:
            $ref: '#/definitions/dto.TaskDto'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 即梦视频接口 (官方兼容)
      tags:
      - Video
  /kling/v1/videos/image2video:
    post:
      consumes:
      - application/json
      description: 调用可灵 AI 图生视频接口提交异步任务。
      parameters:
      - description: 图生视频请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.KlingImage2VideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 任务状态和结果
          schema:
            $ref: '#/definitions/dto.VideoTaskResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 可灵图生视频
      tags:
      - Video
  /kling/v1/videos/image2video/{task_id}:
    get:
      consumes:
      - application/json
      description: 根据 task_id 查询可灵图生视频任务状态。
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务状态和结果
          schema:
            $ref: '#/definitions/dto.VideoTaskResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 可灵任务查询 (图生视频)
      tags:
      - Video
  /kling/v1/videos/text2video:
    post:
      consumes:
      - application/json
      description: 调用可灵 AI 文生视频接口提交异步任务。
      parameters:
      - description: 文生视频请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.KlingText2VideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 任务状态和结果
          schema:
            $ref: '#/definitions/dto.VideoTaskResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 可灵文生视频
      tags:
      - Video
  /kling/v1/videos/text2video/{task_id}:
    get:
      consumes:
      - application/json
      description: 根据 task_id 查询可灵文生视频任务状态。
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务状态和结果
          schema:
            $ref: '#/definitions/dto.VideoTaskResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 可灵任务查询 (文生视频)
      tags:
      - Video
  /v1/video/generations:
    post:
      consumes:
      - application/json
      description: |-
        调用视频生成接口生成视频，支持多种上游：
        - 可灵 AI (Kling): https://app.klingai.com/cn/dev/document-api/apiReference/commonInfo
        - 即梦 (Jimeng): https://www.volcengine.com/docs/85621/1538636
        - 豆包 (Doubao-Seedance): https://www.volcengine.com/docs/82379/1520757
        请求体字段对应 relaycommon.TaskSubmitReq；`metadata` 里的字段会被各渠道 adaptor 按需透传到上游。
      parameters:
      - description: 视频生成请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.VideoGenerationsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 任务对象（Doubao/Sora 等渠道统一 OpenAI Video 对象格式；部分渠道可能返回兼容的 task_id
            包装）
          schema:
            $ref: '#/definitions/dto.OpenAIVideo'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 生成视频
      tags:
      - Video
  /v1/video/generations/{task_id}:
    get:
      consumes:
      - application/json
      description: 根据任务 ID 查询视频生成任务的状态与结果。
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务状态和结果
          schema:
            $ref: '#/definitions/dto.VideoTaskResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 查询视频任务
      tags:
      - Video
  /v1/videos:
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: |-
        OpenAI 兼容的视频生成接口，提交一个异步任务。
        参考：https://platform.openai.com/docs/api-reference/videos/create
        Sora 渠道同时接受 `multipart/form-data`（支持 `input_reference` 文件上传）与 JSON。
      parameters:
      - description: 视频生成请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.VideoGenerationsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 任务对象（OpenAI 视频对象格式）
          schema:
            $ref: '#/definitions/dto.OpenAIVideo'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建视频生成任务 (OpenAI 兼容)
      tags:
      - Video
  /v1/videos/{task_id}:
    get:
      consumes:
      - application/json
      description: 根据 task_id 查询 OpenAI 兼容视频生成任务的状态与结果。
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务对象
          schema:
            $ref: '#/definitions/dto.OpenAIVideo'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 查询视频任务状态 (OpenAI 兼容)
      tags:
      - Video
  /v1/videos/{task_id}/content:
    get:
      description: |-
        根据 task_id 将上游视频内容透传给客户端。支持用户会话或 API token 鉴权。
        成功时响应体为视频二进制流，`Content-Type` 由上游决定（通常为 `video/mp4`）。
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 视频二进制流
          schema:
            type: file
        "400":
          description: task_id 缺失或任务未完成
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "403":
          description: URL 被 SSRF 策略拦截
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "502":
          description: 上游获取视频失败
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 拉取视频内容
      tags:
      - Video
  /v1/videos/{video_id}/remix:
    post:
      consumes:
      - application/json
      description: 基于已有 video_id 的视频提交重混任务。
      parameters:
      - description: 源视频 ID
        in: path
        name: video_id
        required: true
        type: string
      - description: 重混请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.VideoGenerationsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 任务对象
          schema:
            $ref: '#/definitions/dto.OpenAIVideo'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controller.VideoErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 视频重混 (OpenAI 兼容)
      tags:
      - Video
schemes:
- https
- http
securityDefinitions:
  ApiKeyAuth:
    description: 使用 `Bearer sk-xxxxxxx` 的形式在 Header 中传递用户 Token。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"